Torch version SMPL++

+++ source/definition/def.cpp
@@ -42,12 +42,12 @@ namespace smpl {
 
 //===== INTERNAL DEFINITIONS ==================================================
 
-int batch_size = 1;// 1
-int vertex_num = 6890;// 6890
-const int joint_num = 24;// 24
-const int shape_basis_dim = 10;// 10
-const int pose_basis_dim = 207;// 207
-const int face_index_num = 13776;// 13776
+int64_t batch_size = 1;// 1
+int64_t vertex_num = 6890;// 6890
+const int64_t joint_num = 24;// 24
+const int64_t shape_basis_dim = 10;// 10
+const int64_t pose_basis_dim = 207;// 207
+const int64_t face_index_num = 13776;// 13776
 
 //=============================================================================
 } // namespace smpl
